+++ src/components/widgets/Widget.tsx
@@ -1,7 +1,7 @@
 import { SwapSwitch } from 'elements/swapSwitch/SwapSwitch';
 import { ReactComponent as IconTimes } from 'assets/icons/times.svg';
 import { useHistory } from 'react-router-dom';
-import { pools } from 'utils/router';
+import { pools } from 'services/router';
 
 interface Props {
   title: string;
