+++ src/elements/earn/portfolio/liquidityProtection/rewards/MyRewards.tsx
@@ -3,7 +3,7 @@ import { NavLink } from 'react-router-dom';
 import { useMyRewards } from 'elements/earn/portfolio/liquidityProtection/rewards/useMyRewards';
 import { StakeRewardsBtn } from 'elements/earn/portfolio/liquidityProtection/rewards/StakeRewardsBtn';
 import { useAppSelector } from 'redux/index';
-import { portfolioRewardsClaim } from 'utils/router';
+import { portfolioRewardsClaim } from 'services/router';
 
 export const MyRewards = () => {
   const [totalRewards, totalRewardsUsd, claimableRewards, claimableRewardsUsd] =
